+++ src/main/java/company/name/Main.java
@@ -1,12 +1,11 @@
 package company.name;
 
 import company.name.ui.Application;
-import company.name.ui.ApplicationConsoleImpl;
 
 public class Main {
 
     public static void main(String[] args) {
-        Application application = new ApplicationConsoleImpl();
+        Application application = new Application();
         application.run();
     }
